src/utils/helpers.py
import json
import pandas as pd
import yaml
from pathlib import Path

import numpy as np
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
import cartopy.feature as cfeature
import rasterio
import os


def save_or_create(plt, save_path):
    if not os.path.exists(os.path.dirname(save_path)):
        os.makedirs(os.path.dirname(save_path))    
    plt.savefig(save_path)
